requirements.txt
nltk
more-itertools
PorterStemmer
progressbar2
